split digitization model
