+++ SimMuon/GEMDigitizer/interface/GEMDigiModel.h
@@ -7,6 +7,7 @@
  *  Base Class for the GEM strip response simulation 
  *  
  *  \author Sven Dildick
+ *  \modified by Yechan Kang
  */
 
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -17,6 +18,7 @@
 #include "SimDataFormats/TrackerDigiSimLink/interface/StripDigiSimLink.h"
 #include "SimDataFormats/GEMDigiSimLink/interface/GEMDigiSimLink.h"
 #include "SimDataFormats/CrossingFrame/interface/MixCollection.h"
+#include "SimMuon/GEMDigitizer/interface/GEMDigiModule.h"
 
 #include <map>
 #include <set>
@@ -25,59 +27,23 @@ namespace CLHEP {
   class HepRandomEngine;
 }
 
-class GEMEtaPartition;
-class GEMGeometry;
 class PSimHit;
+class GEMEtaPartition;
+class GEMDigiModule;
 
 class GEMDigiModel
 {
 public:
 
-  typedef edm::DetSet<StripDigiSimLink> StripDigiSimLinks;
-  typedef edm::DetSet<GEMDigiSimLink> GEMDigiSimLinks;
-
   virtual ~GEMDigiModel() {}
 
-  void setGeometry(const GEMGeometry *geom) {geometry_ = geom;}
-
-  const GEMGeometry* getGeometry() {return geometry_;}
-
-  virtual void simulateSignal(const GEMEtaPartition*, const edm::PSimHitContainer&, CLHEP::HepRandomEngine* engine) = 0;
-
-  virtual void simulateNoise(const GEMEtaPartition*, CLHEP::HepRandomEngine* engine) = 0;
-  
-  virtual std::vector<std::pair<int,int> > 
-    simulateClustering(const GEMEtaPartition*, const PSimHit*, const int, CLHEP::HepRandomEngine* engine) = 0;
-
-  void fillDigis(int rollDetId, GEMDigiCollection&);
-
-  virtual void setup() = 0;
-
-  const StripDigiSimLinks & stripDigiSimLinks() const {return stripDigiSimLinks_;}
-  const GEMDigiSimLinks & gemDigiSimLinks() const {return theGemDigiSimLinks_;}
+  virtual void simulate(const GEMEtaPartition*, const edm::PSimHitContainer&, CLHEP::HepRandomEngine*) = 0;
 
 protected:
 
-  GEMDigiModel(const edm::ParameterSet&) {}
+  GEMDigiModule* digiModule_;
 
-  const GEMGeometry * geometry_;
-  
-  std::set< std::pair<int, int> > strips_;
-
-  /// creates links from Digi to SimTrack
-  void addLinks(unsigned int strip,int bx);
-  void addLinksWithPartId(unsigned int strip,int bx);
-
-  // keeps track of which hits contribute to which channels
-  typedef std::multimap<
-      std::pair<unsigned int, int>,
-      const PSimHit*,
-      std::less<std::pair<unsigned int, int> >
-    >  DetectorHitMap;
-
-  DetectorHitMap detectorHitMap_;
-  StripDigiSimLinks stripDigiSimLinks_;
-  GEMDigiSimLinks theGemDigiSimLinks_;
+  GEMDigiModel(const edm::ParameterSet&, GEMDigiModule* digiModule) : digiModule_(digiModule) {}
 
 };
 #endif
