+++ SimMuon/GEMDigitizer/src/SealModule.cc
@@ -1,8 +1,10 @@
 #include "FWCore/PluginManager/interface/ModuleDef.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
-#include "SimMuon/GEMDigitizer/interface/GEMDigiModelFactory.h"
 #include "SimMuon/GEMDigitizer/interface/ME0DigiPreRecoModelFactory.h"
 #include "SimMuon/GEMDigitizer/interface/ME0DigiModelFactory.h"
+
+#include "SimMuon/GEMDigitizer/interface/GEMDigiProducer.h"
+DEFINE_FWK_MODULE(GEMDigiProducer);
 
 #include "SimMuon/GEMDigitizer/interface/GEMPadDigiProducer.h"
 DEFINE_FWK_MODULE(GEMPadDigiProducer);
